+++ packages/x-tree-view/src/internals/plugins/useTreeViewSelection/useTreeViewSelection.ts
@@ -1,12 +1,12 @@
 import * as React from 'react';
 import { TreeViewPlugin, TreeViewItemRange } from '../../models';
 import {
-  getNextNavigableItem,
   getFirstNavigableItem,
   getLastNavigableItem,
+  getNavigableItemsInRange,
 } from '../../useTreeView/useTreeView.utils';
 import { UseTreeViewSelectionSignature } from './useTreeViewSelection.types';
-import { convertSelectedItemsToArray, findOrderInTremauxTree } from './useTreeViewSelection.utils';
+import { convertSelectedItemsToArray } from './useTreeViewSelection.utils';
 import { TreeViewItemId } from '../../../models';
 
 export const useTreeViewSelection: TreeViewPlugin<UseTreeViewSelectionSignature> = ({
@@ -93,20 +93,6 @@ export const useTreeViewSelection: TreeViewPlugin<UseTreeViewSelectionSignature>
     currentRangeSelection.current = [];
   };
 
-  const getItemsInRange = (itemAId: string, itemBId: string) => {
-    const [first, last] = findOrderInTremauxTree(instance, itemAId, itemBId);
-    const items = [first];
-
-    let current = first;
-
-    while (current !== last) {
-      current = getNextNavigableItem(instance, current)!;
-      items.push(current);
-    }
-
-    return items;
-  };
-
   const handleRangeArrowSelect = (event: React.SyntheticEvent, items: TreeViewItemRange) => {
     let base = convertSelectedItemsToArray(models.selectedItems.value).slice();
     const { start, end, current } = items;
@@ -147,7 +133,7 @@ export const useTreeViewSelection: TreeViewPlugin<UseTreeViewSelectionSignature>
       base = base.filter((id) => currentRangeSelection.current.indexOf(id) === -1);
     }
 
-    let range = getItemsInRange(start, end);
+    let range = getNavigableItemsInRange(instance, start, end);
     range = range.filter((item) => !instance.isItemDisabled(item));
     currentRangeSelection.current = range;
     let newSelected = base.concat(range);
@@ -193,14 +179,6 @@ export const useTreeViewSelection: TreeViewPlugin<UseTreeViewSelectionSignature>
       start,
       end: getLastNavigableItem(instance),
     });
-  };
-
-  const selectAll = (event: React.SyntheticEvent) => {
-    if (params.disableSelection) {
-      return;
-    }
-
-    const allItems = getItemsInRange();
   };
 
   return {
